Move E2E tests to the Checkout shopper file
