+++ tests/e2e/tests/checkout/checkout-block.shopper.block_theme.side_effects.spec.ts
@@ -3,6 +3,7 @@
  */
 import { expect, test as base } from '@woocommerce/e2e-playwright-utils';
 import { BlockData } from '@woocommerce/e2e-types';
+import { customerFile, guestFile } from '@woocommerce/e2e-utils';
 
 /**
  * Internal dependencies
@@ -10,6 +11,10 @@ import { BlockData } from '@woocommerce/e2e-types';
 import {
 	REGULAR_PRICED_PRODUCT_NAME,
 	SIMPLE_PHYSICAL_PRODUCT_NAME,
+	FREE_SHIPPING_NAME,
+	FREE_SHIPPING_PRICE,
+	FLAT_RATE_SHIPPING_NAME,
+	FLAT_RATE_SHIPPING_PRICE,
 } from './constants';
 import { CheckoutPage } from './checkout.page';
 
@@ -329,5 +334,106 @@ test.describe( 'Shipping and Billing Addresses', () => {
 		await expect(
 			page.getByText( 'Please enter a valid postcode' )
 		).toBeHidden();
+	} );
+} );
+
+test.describe( 'Shopper → Checkout block → Shipping', () => {
+	test.use( { storageState: customerFile } );
+
+	test( 'Shopper can choose free shipping, flat rate shipping, and can have different billing and shipping addresses', async ( {
+		checkoutPageObject,
+		frontendUtils,
+		page,
+	} ) => {
+		await frontendUtils.goToShop();
+		await frontendUtils.emptyCart();
+		await frontendUtils.addToCart( 'Beanie' );
+		await frontendUtils.goToCheckout();
+		await expect(
+			await checkoutPageObject.selectAndVerifyShippingOption(
+				FREE_SHIPPING_NAME,
+				FREE_SHIPPING_PRICE
+			)
+		).toBe( true );
+		await checkoutPageObject.fillInCheckoutWithTestData();
+		await checkoutPageObject.placeOrder();
+		await checkoutPageObject.verifyAddressDetails( 'billing' );
+		await checkoutPageObject.verifyAddressDetails( 'shipping' );
+		await expect( page.getByText( FREE_SHIPPING_NAME ) ).toBeVisible();
+
+		await frontendUtils.goToShop();
+		await frontendUtils.addToCart( 'Beanie' );
+		await frontendUtils.goToCheckout();
+		await expect(
+			await checkoutPageObject.selectAndVerifyShippingOption(
+				FLAT_RATE_SHIPPING_NAME,
+				FLAT_RATE_SHIPPING_PRICE
+			)
+		).toBe( true );
+
+		await checkoutPageObject.syncBillingWithShipping();
+		await checkoutPageObject.fillInCheckoutWithTestData( {
+			phone: '0987654322',
+		} );
+		await checkoutPageObject.unsyncBillingWithShipping();
+		const shippingForm = page.getByRole( 'group', {
+			name: 'Shipping address',
+		} );
+		const billingForm = page.getByRole( 'group', {
+			name: 'Billing address',
+		} );
+
+		await expect( shippingForm.getByLabel( 'Phone' ).inputValue ).toEqual(
+			billingForm.getByLabel( 'Phone' ).inputValue
+		);
+
+		await checkoutPageObject.fillInCheckoutWithTestData();
+		const overrideBillingDetails = {
+			firstname: 'Juan',
+			lastname: 'Perez',
+			addressfirstline: '123 Test Street',
+			addresssecondline: 'Apartment 6',
+			country: 'ES',
+			city: 'Madrid',
+			postcode: '08830',
+			state: 'M',
+			phone: '0987654321',
+			email: 'juan.perez@test.com',
+		};
+		await checkoutPageObject.fillBillingDetails( overrideBillingDetails );
+		await checkoutPageObject.placeOrder();
+		await checkoutPageObject.verifyAddressDetails(
+			'billing',
+			overrideBillingDetails
+		);
+		await checkoutPageObject.verifyAddressDetails( 'shipping' );
+	} );
+} );
+
+// We only check if guest user can place an order because we already checked if logged in user can
+// place an order in the previous test
+test.describe( 'Shopper → Checkout block → Place Order', () => {
+	test.use( { storageState: guestFile } );
+
+	test( 'Guest user can place order', async ( {
+		checkoutPageObject,
+		frontendUtils,
+		page,
+	} ) => {
+		await frontendUtils.emptyCart();
+		await frontendUtils.goToShop();
+		await frontendUtils.addToCart( SIMPLE_PHYSICAL_PRODUCT_NAME );
+		await frontendUtils.goToCheckout();
+		await expect(
+			await checkoutPageObject.selectAndVerifyShippingOption(
+				FREE_SHIPPING_NAME,
+				FREE_SHIPPING_PRICE
+			)
+		).toBe( true );
+		await checkoutPageObject.fillInCheckoutWithTestData();
+		await checkoutPageObject.placeOrder();
+		await expect(
+			page.getByText( 'Your order has been received.' )
+		).toBeVisible();
 	} );
 } );
